split to file
